refactor: move src/schemas to src/controllers/schemas
